requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
telebot==0.0.5
urllib3==2.0.4
DateTime==5.2

